added the about page

+++ migrationdata/pyflask/main.py
@@ -73,15 +73,17 @@ def plotgraph():
                                             ')')]
     tempdf = tempdf[tempdf['citizenship'].apply(lambda x: x not in set(['All', 'Locals']))]
     fig, ax = plt.subplots(figsize=(9.5, 6))
-    tempdf[tempdf['mau_audience'] > 1000].set_index('citizenship').sort_values('mau_audience', ascending=False).head(
-        10)[::-1]['mau_audience'].plot(kind='barh')
-    ax.set_xlabel('', labelpad=15)
-    ax.set_ylabel('', labelpad=30)
-    plt.savefig('static/plot.png', transparent=True)
-    encoded = base64.b64encode(open('static/plot.png', 'rb').read()).decode()
-    plothtml = 'data:image/png;base64,{}'
-    plothtml = plothtml.format(encoded)
-    os.remove('static/plot.png')
+    if tempdf[tempdf['mau_audience'] > 1000].empty:
+        plothtml = ""
+    else:
+        tempdf[tempdf['mau_audience'] > 1000].set_index('citizenship').sort_values('mau_audience', ascending=False).head(10)[::-1]['mau_audience'].plot(kind='barh')
+        ax.set_xlabel('', labelpad=15)
+        ax.set_ylabel('', labelpad=30)
+        plt.savefig('static/plot.png', transparent=True)
+        encoded = base64.b64encode(open('static/plot.png', 'rb').read()).decode()
+        plothtml = 'data:image/png;base64,{}'
+        plothtml = plothtml.format(encoded)
+        os.remove('static/plot.png')
 
     return render_template('plotgraph.html', plot=plothtml, gender=gender, scholarities=scholarities, os=os_var)
 
@@ -101,15 +103,17 @@ def country(countrycode):
     df = pd.read_csv(path)
     df = df[df['citizenship'].apply(lambda x: x not in set(['All', 'Locals']))]
 
+    print(df.columns)
     fig, ax = plt.subplots(figsize=(9.5, 6))
     df[df['Total_mau']>1000].set_index('citizenship').sort_values('Total_mau', ascending=False).head(10)['Total_mau'][::-1].plot(kind='barh')
+
     ax.set_xlabel('', labelpad=15)
     ax.set_ylabel('', labelpad=30)
 
-    plt.savefig('static/plotcountry{}.png'.format(countrycode), transparent=True)
-    encoded = base64.b64encode(open('static/plotcountry{}.png'.format(countrycode), 'rb').read()).decode()
+    plt.savefig('static/plotcountry1{}.png'.format(countrycode), transparent=True)
+    encoded = base64.b64encode(open('static/plotcountry1{}.png'.format(countrycode), 'rb').read()).decode()
     html1 = 'data:image/png;base64,{}'.format(encoded)
-    os.remove('static/plotcountry{}.png'.format(countrycode))
+    os.remove('static/plotcountry1{}.png'.format(countrycode))
 
     url = 'http://data.un.org/en/iso/{}.html'.format(countrycode)
     countryData = requests.get(url).text
@@ -246,5 +250,8 @@ def Qatar2():
     return render_template("Country.html")
 
 
+@app.route('/about')
+def about():
+    return render_template("about.html")
 if __name__ == "__main__":
     app.run(debug=True)
